+++ releases.md
@@ -1,5 +1,7 @@
 # `greek` library: release notes
 
+**7.0.0**: completes refactoring begun with 6.0.0 release.  Transcoding of ascii Strings and unicode
+code points is handled by the `CodePointTranscoder` object.  The `LiteraryGreekString` now extends the `LGSTrait` as well as the `GreekString` trait.  The `LiteraryGreekOrthography` defines character classes used by the `LiteraryGreekString` class, and implements the `MidOrthography` trait.
 
 **6.0.0**: refactor `LiteraryGreekString` as an implementation of a new `LGSTrait`.